requirements.txt
pillow==12.0.0
pyyaml==6.0.3
